Clean up LSPServer and use byte rather than string length for messages - fixes #736
