+++ packages/@romefrontend/core/server/lsp/protocol.test.md
@@ -1,0 +1,25 @@
+# `protocol.test.ts`
+
+**DO NOT MODIFY**. This file has been autogenerated. Run `rome test packages/@romefrontend/core/server/lsp/protocol.test.ts --update-snapshots` to update.
+
+## `LSPTransport handles emoji byte lengths`
+
+```
+⚡ [LSPServer][Transport] Status: WAITING_FOR_HEADERS_END
+⚡ [LSPServer][Transport] Headers for next message: Object {
+    expectedLength: 321
+    extra: Map []
+  }
+⚡ [LSPServer][Transport] Status: WAITING_FOR_RESPONSE_END
+⚡ [LSPServer][Transport] Received message content:
+  {"jsonrpc":"2.0","method":"textDocument/didChange","params":{"textDocument":{"uri":"file:///Users/s
+  ebastianmckenzie/Scratch/rome-playground/test.ts","version":24},"contentChanges":[{"text":"//Can’t
+  handle this comment
+  const foo = 'Or this “special” string';
+  const rocket = "Or this🚀";
+  
+  rocket;
+  foo;"}]}}
+⚡ [LSPServer][Transport] Status: IDLE
+
+```
