+++ scripts/run-vscode-extension
@@ -15,9 +15,7 @@ const os = require("os");
 const outFolder = path.join(os.tmpdir(), "rome-vscode-dev");
 
 async function main() {
-  await buildTrunk();
-
-  heading("Bundling Extension");
+  heading("Bundling extension");
   await execDev(["bundle", "@romefrontend-integration/vscode", outFolder]);
 
   heading("Running VSCode");
